Add new object LocalPrefixSpan, and do some optimization.
